automation/requirements.txt
autopep8==1.2.4
backports-abc==0.4
blitzdb==0.2.12
gitdb==0.6.4
GitPython==2.0.2
nose==1.3.7
pep8==1.7.0
PyGithub==1.26.0
six==1.10.0
smmap==0.9.0
